Copia de Project Diario Oficial/requirements.txt
# Framework Web
Django==5.0.6
gunicorn

# Base de Datos
psycopg2-binary

# Tareas en segundo plano
celery
redis

# Cache
django-redis

# Scraper y procesamiento
requests
beautifulsoup4
selenium
webdriver-manager
pandas
openpyxl
Pillow
pypdf
pdf2image
pytesseract
google-generativeai
pdfminer.six
PyPDF2

# Variables de entorno
python-dotenv

# API de WhatsApp
twilio

# Added from the code block
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
lxml
schedule
dj-database-url
whitenoise[brotli]
honcho
undetected-chromedriver